Add minimal lv size support
The minimum size to extend the logical volume up to. If actual size is greater, then do nothing.

This use case for this was to extend several machines that /var was not standard everywhere and it was required to extend the size up to a specific size only where it was lower than the minimal size.